[Chore] `--security` for snapshot and OSD server (#5451)
Add the ability to run:
```
yarn opensearch snapshot --security
```
which will start the OpenSearch cluster with the security plugin.

And add the ability to run:
```
yarn start:security
```
which will start the OpenSearch Dashboards server if the security
plugin is available.

This is only intended for demo and local purposes.

Issue:
n/a

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>